+++ sctplib/programs/main.c
@@ -1,5 +1,5 @@
 /*
- *  $Id: main.c,v 1.5 2004/11/19 21:25:00 tuexen Exp $
+ *  $Id: main.c,v 1.6 2005/03/08 13:29:52 tuexen Exp $
  *
  * SCTP implementation according to RFC 2960.
  * Copyright (C) 2000 by Siemens AG, Munich, Germany.
@@ -79,7 +79,7 @@ gboolean use_unordered = FALSE;
 
 /* this one is global, as we want to know, whether to enable ipv6 calls, even if it may be compiled in */
 gboolean support_ipv6;
-short sctp_instance,sctp_instance2,sctp_instance3;
+int sctp_instance,sctp_instance2,sctp_instance3;
 unsigned short ulpPort = 0;
